+++ test/manipulate.jl
@@ -21,8 +21,8 @@ sharks = @gulfsharks;
     @testset "decimal-minutes locations" begin
         x = fill("11 22.33W", length(samplenames(cats))) ; y = fill("-41 31.52", length(samplenames(cats)))
         locationdata!(cats, longitude = x, latitude = y)
-        @test all(cats.sampleinfo.longitude .== Float32(-11.3722))
-        @test all(cats.sampleinfo.latitude .== Float32(-41.5253))
+        @test all(cats.sampleinfo.longitude .== -11.3722)
+        @test all(cats.sampleinfo.latitude .== -41.5253)
     end
 
     @testset "loci and genotypes" begin
